Refactor bonus calculation
